requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.6
adb-shell>=0.4.3
cryptography>=36.0.0 